[frontend] add 'me' to list of users in filters
