+++ opencti-platform/opencti-front/src/private/components/common/form/ObjectAssigneeField.jsx
@@ -37,6 +37,11 @@ export const objectAssigneeFieldAssigneesSearchQuery = graphql`
         }
       }
     }
+    me {
+      id
+      entity_type
+      name
+    }
   }
 `;
