+++ opencti-platform/opencti-front/src/private/components/common/form/ObjectParticipantField.tsx
@@ -35,6 +35,11 @@ export const objectParticipantFieldParticipantsSearchQuery = graphql`
         }
       }
     }
+    me {
+      id
+      entity_type
+      name
+    }
   }
 `;
