+++ opencti-platform/opencti-front/src/private/components/common/identities/IdentitySearch.js
@@ -33,5 +33,10 @@ export const identitySearchCreatorsSearchQuery = graphql`
         }
       }
     }
+    me {
+      id
+      entity_type
+      name
+    }
   }
 `;
